Remove colortemplate dependency
In https://github.com/lifepillar/vim-colortemplate/issues/43, the
maintainers of colortemplate are discussing a new version. From the
discussion it's becoming clear that colortemplate isn't interested in
long term support for NeoVim.

Since colortemplate wasn't _that_ useful, this commit rewrites the Vim
template file to rely on nothing but Vim itself. To make this process a
bit easier we use a custom (temporary) command, removed at the end of
the color file.

As part of this change, we no longer make any attempt at supporting
anything but true color GUIs. This removes the need for maintaining
different color values, and most modern (Neo)Vim GUIs (even when used in
a modern terminal) should support true colors.