Make statuslines look the same
Using different colors for StatusLine and StatusLineNC doesn't work very
well when using a light background. Most notably, it becomes difficult
to highlight the active statusline without this becoming too
distracting.

To work around this, StatusLine and StatusLineNC now look the same. To
highlight the active statusline/window, it's best to highlight specific
elements (such as the file) differently.